pyproject.toml
[tool.poetry]
name = "portfolio"
version = "0.1.0"
description = ""
authors = ["Vitalii-Kh95 <mew030995@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5"
djangorestframework = "*"
django-cors-headers = "*"
django-taggit = "^6.1.0"
pillow = "^11.0.0"


[tool.poetry.group.dev.dependencies]
faker = "^30.8.1"
pytest = "^8.3.4"
httpx = "^0.28.1"
pandas = "^2.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
